Adjust keycloak helm chart to avoid troubles with routing.
I did verify Keycloak configuration and 19.0.3 (.2 probably too) have a bug related to detection of https from actual configuration.
There are few options which are available the most stable seem to be HOSTNAME_URL which allows to pass https, host and port all together.
The https detection with hostname alone doesn't work and needs KC_HOSTNAME_STRICT_HTTPS to be set to false, which is undocumented feature.

Shipping it for now, as there is a pending issue in Keycloak itself: keycloak/keycloak#15287, which awaits resolution in release later than we use.

Signed-off-by: Łukasz Dywicki <luke@code-house.org>