HS-1287: Persistence for enqueued sink messages
Outgoing messages are aggregated and serialized before the enqueing.
Therefore the queue is type independent and shared between sink modules.

The queue now supports off-heap storage of elements which is also used
for persistence on restarts. The queue maintains two counters shared for
all modules. First counter is the number of elements in memory, second
is the number of total elements in queue. Whenever a new element is
enqued, the element is kept in memory. If required, the oldest element
in memory is persisted to provide more room. Dequeing will try to
consume elements from off-heap storage first, followed by the in-memory
elements.